+++ Makefile
@@ -1,17 +1,17 @@
 # Copyright 2021 Absa Group Limited
-# 
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
+#
 #     http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-# 
+#
 # Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 .PHONY: lint
 lint:
@@ -19,7 +19,7 @@ lint:
 
 .PHONY: license
 license:
-	go install github.com/AbsaOSS/golic@v0.1.2
+	go install github.com/AbsaOSS/golic@v0.6.0
 	golic inject -c="2021 Absa Group Limited"
 
 .PHONY: test
