+++ reflect/reflect.go
@@ -21,6 +21,9 @@ import "reflect"
 
 func GetType(v interface{}) string {
 	t := reflect.TypeOf(v)
+	if t == nil {
+		return "nil"
+	}
 	if t.Kind() == reflect.Ptr {
 		return "*" + t.Elem().Name()
 	}
