+++ reflect/reflect_test.go
@@ -15,14 +15,47 @@ limitations under the License.
 
 Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 */
-package shell
+package reflect
 
-// Command is a synchronized wrapper around the standard Go command, which allows you to define environment variables
-// in a native way and returns the output as a return value
-type Command struct {
-	Command           string            // The command to run
-	Args              []string          // The args to pass to the command
-	WorkingDir        string            // The working directory
-	Env               map[string]string // Additional environment variables to set
-	OutputMaxLineSize int               // The max line size of stdout and stderr (in bytes)
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGetTypeStruct(t *testing.T) {
+	// arrange
+	type ab struct{}
+	x := ab{}
+	// act
+	r := GetType(x)
+	// assert
+	assert.Equal(t, "ab", r)
+}
+
+func TestGetTypePrimitive(t *testing.T) {
+	// arrange
+	x := 5
+	// act
+	r := GetType(x)
+	// assert
+	assert.Equal(t, "int", r)
+}
+
+func TestGetTypePointer(t *testing.T) {
+	// arrange
+	type ab struct{}
+	x := &ab{}
+	// act
+	r := GetType(x)
+	// assert
+	assert.Equal(t, "*ab", r)
+}
+
+func TestGetTypeNil(t *testing.T) {
+	// arrange
+	// act
+	r := GetType(nil)
+	// assert
+	assert.Equal(t, "nil", r)
 }
