--- reflect/reflection_test.go
@@ -1,53 +1,0 @@
-/*
-Copyright 2021 Absa Group Limited
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
-Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
-*/
-package reflect
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestGetTypeStruct(t *testing.T) {
-	// arrange
-	type ab struct{}
-	x := ab{}
-	// act
-	r := GetType(x)
-	// assert
-	assert.Equal(t, "ab", r)
-}
-
-func TestGetTypePrimitive(t *testing.T) {
-	// arrange
-	x := 5
-	// act
-	r := GetType(x)
-	// assert
-	assert.Equal(t, "int", r)
-}
-
-func TestGetTypePointer(t *testing.T) {
-	// arrange
-	type ab struct{}
-	x := &ab{}
-	// act
-	r := GetType(x)
-	// assert
-	assert.Equal(t, "*ab", r)
-}
