+++ string/string.go
@@ -16,7 +16,7 @@ limitations under the License.
 Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 */
 // Package strings provides common extensions for string parsing and formatting
-package strings
+package string
 
 import (
 	"encoding/json"
